fix: ensure system integrity with best-effort briefing and singleton sentinel deduplication
